Change context strategy to serial background saving
Ditched multple background context that saved changes in parallel. This gives to much pain with deletions and faults that can't be fetched.

We now use a backgroundContext that performs all changes serially (still async).

Guideline:
- Main context is for UI blocking READING ONLY, for example NSFetchedResultsController
- Root context should not be used directly
- Background context should be used for ALL other READS/WRITES

You can spawn extra child contexts yourself on any of these contexts if you know what you're doing. Usefull for very large reads that shouldn't block other operations. But be aware of the risks that we try to avoid with this fix.